Add STM32 breakout board. Update schematics PDF and docs.
The PDF wasn't updated with the previous commit, but should have been. Likewise the documentation.
The specific issue was the current measurement circuitry needed a direct connection to the Common terminal.

The STM32 breakout board is for prototyping. Any QFP64 breakout boards I've seen have pins (or holes)
on four sides, which is unsuitable for a breadboard. This is a simple breakout board if the net ties
are cut and no components are placed. Otherwise it is actually a very minimal STM32F373 dev board.